Reuse managers from base in android
