--- android/sdl_android/src/main/java/com/smartdevicelink/managers/CompletionListener.java
@@ -1,10 +1,0 @@
-package com.smartdevicelink.managers;
-
-public interface CompletionListener {
-
-	/**
-	 * Returns whether a specific operation was successful or not
-	 * @param success - success or fail
-	 */
-	void onComplete(boolean success);
-}
