+++ base_android/src/main/java/com/smartdevicelink/managers/StreamingStateMachine.java
@@ -1,0 +1,1 @@
+../../../../../../../base/src/main/java/com/smartdevicelink/managers/StreamingStateMachine.java
