+++ base_android/src/main/java/com/smartdevicelink/managers/file
@@ -1,0 +1,1 @@
+../../../../../../../base/src/main/java/com/smartdevicelink/managers/file
