+++ base_android/src/main/java/com/smartdevicelink/managers/screen
@@ -1,0 +1,1 @@
+../../../../../../../base/src/main/java/com/smartdevicelink/managers/screen
