+++ base_android/src/main/java/com/smartdevicelink/managers/video
@@ -1,0 +1,1 @@
+../../../../../../../base/src/main/java/com/smartdevicelink/managers/video/
