+++ base_android/src/main/java/com/smartdevicelink/protocol/enums
@@ -1,0 +1,1 @@
+../../../../../../../base/src/main/java/com/smartdevicelink/protocol/enums
